feat: implement hightide version 0.1.24
