+++ api-services/mutations/properties/property_view_src_mutations.ts
@@ -1,54 +1,25 @@
+import type { UseMutationOptions } from '@tanstack/react-query'
 import { useMutation, useQueryClient } from '@tanstack/react-query'
-import {
-  FilterUpdate,
-  UpdatePropertyViewRuleRequest
-} from '@helpwave/proto-ts/services/property_svc/v1/property_views_svc_pb'
-import {
-  PatientPropertyMatcher,
-  TaskPropertyMatcher
-} from '@helpwave/proto-ts/services/property_svc/v1/property_value_svc_pb'
 import { QueryKeys } from '../query_keys'
-import type { SubjectType } from '../../types/properties/property'
-import { APIServices } from '../../services'
-import { getAuthenticatedGrpcMetadata } from '../../authentication/grpc_metadata'
+import type { PropertySubjectType } from '../../types/properties/property'
+import type {
+  PropertyViewRuleFilterUpdate
+} from '../../service/properties/PropertyViewSourceService'
+import {
+  PropertyViewSourceService
+} from '../../service/properties/PropertyViewSourceService'
 
-type PropertyViewRuleFilterUpdate = {
-  subjectId: string,
-  appendToAlwaysInclude?: string[],
-  removeFromAlwaysInclude?: string[],
-  appendToDontAlwaysInclude?: string[],
-  removeFromDontAlwaysInclude?: string[],
-}
-
-export const useUpdatePropertyViewRuleRequest = (subjectType: SubjectType, wardId?: string) => {
+export const useUpdatePropertyViewRuleRequest = (subjectType: PropertySubjectType, wardId?: string, options?: UseMutationOptions<boolean, unknown, PropertyViewRuleFilterUpdate>) => {
   const queryClient = useQueryClient()
   return useMutation({
+    ...options,
     mutationFn: async (update: PropertyViewRuleFilterUpdate) => {
-      const req = new UpdatePropertyViewRuleRequest()
-      if (subjectType === 'patient') {
-        const matcher = new PatientPropertyMatcher().setPatientId(update.subjectId)
-        if (wardId) {
-          matcher.setWardId(wardId)
-        }
-        req.setPatientMatcher(matcher)
-      }
-      if (subjectType === 'task') {
-        const matcher = new TaskPropertyMatcher().setTaskId(update.subjectId)
-        if (wardId) {
-          matcher.setWardId(wardId)
-        }
-        req.setTaskMatcher(matcher)
-      }
-
-      req.setFilterUpdate(new FilterUpdate()
-        .setAppendToAlwaysIncludeList(update.appendToAlwaysInclude ?? [])
-        .setRemoveFromAlwaysIncludeList(update.removeFromAlwaysInclude ?? [])
-        .setAppendToDontAlwaysIncludeList(update.removeFromAlwaysInclude ?? [])
-        .setRemoveFromDontAlwaysIncludeList(update.removeFromDontAlwaysInclude ?? []))
-
-      await APIServices.propertyViewSource.updatePropertyViewRule(req, getAuthenticatedGrpcMetadata())
+      return await PropertyViewSourceService.update(update, subjectType, wardId)
     },
-    onSuccess: () => {
+    onSuccess: (data, variables, context) => {
+      if(options?.onSuccess) {
+        options.onSuccess(data, variables, context)
+      }
       queryClient.invalidateQueries([QueryKeys.properties]).catch(console.error)
     }
   })
