+++ api-services/mutations/tasks/bed_mutations.ts
@@ -1,54 +1,32 @@
+import type { UseMutationOptions } from '@tanstack/react-query'
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query'
-import {
-  CreateBedRequest,
-  DeleteBedRequest,
-  GetBedRequest,
-  UpdateBedRequest
-} from '@helpwave/proto-ts/services/tasks_svc/v1/bed_svc_pb'
 import { QueryKeys } from '../query_keys'
 import { APIServices } from '../../services'
-import { getAuthenticatedGrpcMetadata } from '../../authentication/grpc_metadata'
 import type { BedWithRoomId } from '../../types/tasks/bed'
 import { roomOverviewsQueryKey } from './room_mutations'
+import { BedService } from '../../service/tasks/BedService'
 
-export const useBedQuery = (bedId: string | undefined) => {
+export const useBedQuery = (id?: string) => {
   return useQuery({
     queryKey: [QueryKeys.beds],
-    enabled: !!bedId,
+    enabled: !!id,
     queryFn: async () => {
-      const req = new GetBedRequest()
-      if (bedId) {
-        req.setId(bedId)
-      }
-      const res = await APIServices.bed.getBed(req, getAuthenticatedGrpcMetadata())
-
-      const bed: BedWithRoomId = {
-        id: res.getId(),
-        name: res.getName(),
-        roomId: res.getRoomId()
-      }
-
-      return bed
+      return await BedService.get(id!)
     },
   })
 }
 
-export const useBedCreateMutation = () => {
+export const useBedCreateMutation = (options?: UseMutationOptions<BedWithRoomId, unknown, BedWithRoomId>) => {
   const queryClient = useQueryClient()
   return useMutation({
+    ...options,
     mutationFn: async (bed: BedWithRoomId) => {
-      const req = new CreateBedRequest()
-      req.setRoomId(bed.roomId)
-      req.setName(bed.name)
-      const res = await APIServices.bed.createBed(req, getAuthenticatedGrpcMetadata())
-
-      if (!res.toObject()) {
-        console.error('error in BedCreate')
-      }
-
-      return { id: res.getId(), name: bed.name }
+      return await BedService.create(bed)
     },
-    onSuccess: () => {
+    onSuccess: (data, variables, context) => {
+      if (options?.onSuccess) {
+        options.onSuccess(data, variables, context)
+      }
       queryClient.invalidateQueries([QueryKeys.beds]).catch(console.error)
       queryClient.invalidateQueries([QueryKeys.rooms, roomOverviewsQueryKey]).catch(console.error)
       queryClient.invalidateQueries([QueryKeys.wards]).catch(console.error)
@@ -56,50 +34,34 @@ export const useBedCreateMutation = () => {
   })
 }
 
-export const useBedUpdateMutation = () => {
+export const useBedUpdateMutation = (options?: UseMutationOptions<boolean, unknown, BedWithRoomId>) => {
   const queryClient = useQueryClient()
   return useMutation({
+    ...options,
     mutationFn: async (bed: BedWithRoomId) => {
-      const req = new UpdateBedRequest()
-      req.setId(bed.id)
-      req.setName(bed.name)
-      req.setRoomId(bed.roomId)
-
-      const res = await APIServices.bed.updateBed(req, getAuthenticatedGrpcMetadata())
-
-      const obj = res.toObject() // TODO: what is the type of this?
-
-      if (!obj) {
-        throw new Error('error in BedUpdate')
-      }
-
-      return obj
+      return await BedService.update(bed)
     },
-    onSuccess: () => {
+    onSuccess: (data, variables, context) => {
+      if (options?.onSuccess) {
+        options.onSuccess(data, variables, context)
+      }
       queryClient.invalidateQueries([APIServices.bed]).catch(console.error)
       queryClient.invalidateQueries([QueryKeys.rooms, roomOverviewsQueryKey]).catch(console.error)
     },
   })
 }
 
-export const useBedDeleteMutation = () => {
+export const useBedDeleteMutation = (options?: UseMutationOptions<boolean, unknown, string>) => {
   const queryClient = useQueryClient()
   return useMutation({
+    ...options,
     mutationFn: async (bedId: string) => {
-      const req = new DeleteBedRequest()
-      req.setId(bedId)
-
-      const res = await APIServices.bed.deleteBed(req, getAuthenticatedGrpcMetadata())
-
-      const obj = res.toObject() // TODO: what is the type of this?
-
-      if (!obj) {
-        throw new Error('error in BedDelete')
-      }
-
-      return obj
+      return await BedService.delete(bedId)
     },
-    onSuccess: () => {
+    onSuccess: (data, variables, context) => {
+      if (options?.onSuccess) {
+        options.onSuccess(data, variables, context)
+      }
       queryClient.invalidateQueries([APIServices.bed]).catch(console.error)
       queryClient.invalidateQueries([QueryKeys.rooms, roomOverviewsQueryKey]).catch(console.error)
       queryClient.invalidateQueries([QueryKeys.wards]).catch(console.error)
