+++ tasks/components/OrganizationInvitationList.tsx
@@ -123,7 +123,6 @@ export const OrganizationInvitationList = ({
       isLoading={isLoading && !!context.state.organizationId}
       hasError={isError && !!context.state.organizationId}
       className="min-h-72"
-      minimumLoadingDuration={200}
     >
       <InputModal
         className="min-w-[400px]"
