+++ tasks/components/cards/OrganizationCard.tsx
@@ -1,4 +1,3 @@
-
 import { Mail } from 'lucide-react'
 import type { Translation } from '@helpwave/hightide'
 import { useTranslation, type PropsForTranslation } from '@helpwave/hightide'
@@ -31,10 +30,10 @@ export type OrganizationCardProps = EditCardProps & {
  * A Card displaying a Organization
  */
 export const OrganizationCard = ({
-  overwriteTranslation,
-  organization,
-  ...editCardProps
-}: PropsForTranslation<OrganizationCardTranslation, OrganizationCardProps>) => {
+                                   overwriteTranslation,
+                                   organization,
+                                   ...editCardProps
+                                 }: PropsForTranslation<OrganizationCardTranslation, OrganizationCardProps>) => {
   const translation = useTranslation([defaultOrganizationCardTranslation], overwriteTranslation)
   const organizationMemberCount = organization.members.length
 
@@ -55,7 +54,13 @@ export const OrganizationCard = ({
           <div className="text-left my-1 font-semibold text-gray-600 text-sm truncate">
             {`${organizationMemberCount} ${organizationMemberCount > 1 ? translation('members') : translation('member')}`}
           </div>
-          <AvatarGroup avatars={organization.members.map(user => ({ avatarUrl: user.avatarURL, alt: user.name }))}/>
+          <AvatarGroup
+            avatars={organization.members.map(user => ({
+              image: { avatarUrl: user.avatarURL, alt: user.name },
+              name: user.name,
+            }))}
+            fullyRounded={true}
+          />
         </div>
       </div>
     </EditCard>
