+++ tasks/package.json
@@ -14,7 +14,7 @@
     "@dnd-kit/core": "6.3.1",
     "@dnd-kit/sortable": "7.0.2",
     "@helpwave/api-services": "workspace:*",
-    "@helpwave/hightide": "^0.1.21",
+    "@helpwave/hightide": "^0.1.24",
     "@tailwindcss/postcss": "^4.1.3",
     "@tanstack/react-query": "^4.36.1",
     "@tanstack/react-query-devtools": "^4.36.1",
